requirements.txt
apple-music-python==1.0.4
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2
django-crontab==0.7.1
django-environ==0.4.5
ecdsa==0.14.1
fcm-django==0.3.10
gcloud==0.17.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
jws==0.1.3
oauth2client==3.0.0
protobuf==3.16.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pyfcm==1.4.9
PyJWT==2.0.1
pyparsing==2.4.7
python-jose==3.2.0
python-jwt==2.0.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.15.0
spotipy==2.18.0
sqlparse==0.4.1
urllib3==1.26.4
validate-email==1.3
